Renamed package and driver name from gomethods to generic
